Move runtime macro inside its own module.
